refactor: can now use flutter_launcher_icons key in pubspec instead of flutter_icons
