+++ lib/main.dart
@@ -8,7 +8,7 @@ import 'package:flutter_launcher_icons/android.dart' as android_launcher_icons;
 import 'package:flutter_launcher_icons/constants.dart' as constants;
 import 'package:flutter_launcher_icons/constants.dart';
 import 'package:flutter_launcher_icons/custom_exceptions.dart';
-import 'package:flutter_launcher_icons/flutter_launcher_icons_config.dart';
+import 'package:flutter_launcher_icons/config/config.dart';
 import 'package:flutter_launcher_icons/ios.dart' as ios_launcher_icons;
 import 'package:flutter_launcher_icons/logger.dart';
 import 'package:flutter_launcher_icons/macos/macos_icon_generator.dart';
@@ -104,7 +104,7 @@ Future<void> createIconsFromArguments(List<String> arguments) async {
       for (String flavor in flavors) {
         print('\nFlavor: $flavor');
         final flutterLauncherIconsConfigs =
-            FlutterLauncherIconsConfig.loadConfigFromFlavor(flavor, prefixPath);
+            Config.loadConfigFromFlavor(flavor, prefixPath);
         if (flutterLauncherIconsConfigs == null) {
           throw NoConfigFoundException(
             'No configuration found for $flavor flavor.',
@@ -127,7 +127,7 @@ Future<void> createIconsFromArguments(List<String> arguments) async {
 }
 
 Future<void> createIconsFromConfig(
-  FlutterLauncherIconsConfig flutterConfigs,
+  Config flutterConfigs,
   FLILogger logger,
   String prefixPath, [
   String? flavor,
@@ -168,15 +168,15 @@ Future<void> createIconsFromConfig(
   );
 }
 
-FlutterLauncherIconsConfig? loadConfigFileFromArgResults(
+Config? loadConfigFileFromArgResults(
   ArgResults argResults,
 ) {
   final String prefixPath = argResults[prefixOption];
   final flutterLauncherIconsConfigs =
-      FlutterLauncherIconsConfig.loadConfigFromPath(
+      Config.loadConfigFromPath(
             argResults[fileOption],
             prefixPath,
           ) ??
-          FlutterLauncherIconsConfig.loadConfigFromPubSpec(prefixPath);
+          Config.loadConfigFromPubSpec(prefixPath);
   return flutterLauncherIconsConfigs;
 }
