+++ lib/src/version.dart
@@ -1,2 +1,2 @@
 // Generated code. Do not modify.
-const packageVersion = '0.13.0';
+const packageVersion = '0.13.1';
