+++ test/android_test.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:flutter_launcher_icons/android.dart' as android;
 import 'package:flutter_launcher_icons/constants.dart';
-import 'package:flutter_launcher_icons/flutter_launcher_icons_config.dart';
+import 'package:flutter_launcher_icons/config/config.dart';
 import 'package:test/test.dart';
 
 // unit tests for android.dart
@@ -35,7 +35,7 @@ void main() {
       'ios': true
     };
     expect(
-      FlutterLauncherIconsConfig.fromJson(flutterIconsConfig)
+      Config.fromJson(flutterIconsConfig)
           .isCustomAndroidFile,
       isFalse,
     );
@@ -46,7 +46,7 @@ void main() {
       'ios': true
     };
     expect(
-      FlutterLauncherIconsConfig.fromJson(flutterIconsNewIconConfig)
+      Config.fromJson(flutterIconsNewIconConfig)
           .isCustomAndroidFile,
       isTrue,
     );
@@ -60,7 +60,7 @@ void main() {
       'ios': true
     };
     expect(
-      FlutterLauncherIconsConfig.fromJson(flutterIconsNewIconConfig)
+      Config.fromJson(flutterIconsNewIconConfig)
           .getImagePathAndroid(),
       equals('assets/images/icon-android.png'),
     );
