+++ test/windows/windows_icon_generator_test.mocks.dart
@@ -4,9 +4,9 @@
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:cli_util/cli_logging.dart' as _i2;
-import 'package:flutter_launcher_icons/flutter_launcher_icons_config.dart'
-    as _i3;
-import 'package:flutter_launcher_icons/logger.dart' as _i4;
+import 'package:flutter_launcher_icons/config/config.dart' as _i3;
+import 'package:flutter_launcher_icons/config/windows_config.dart' as _i4;
+import 'package:flutter_launcher_icons/logger.dart' as _i5;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -40,12 +40,11 @@ class _FakeProgress_1 extends _i1.SmartFake implements _i2.Progress {
         );
 }
 
-/// A class which mocks [FlutterLauncherIconsConfig].
+/// A class which mocks [Config].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockFlutterLauncherIconsConfig extends _i1.Mock
-    implements _i3.FlutterLauncherIconsConfig {
-  MockFlutterLauncherIconsConfig() {
+class MockConfig extends _i1.Mock implements _i3.Config {
+  MockConfig() {
     _i1.throwOnMissingStub(this);
   }
 
@@ -67,6 +66,21 @@ class MockFlutterLauncherIconsConfig extends _i1.Mock
   @override
   bool get hasPlatformConfig => (super.noSuchMethod(
         Invocation.getter(#hasPlatformConfig),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get hasWebConfig => (super.noSuchMethod(
+        Invocation.getter(#hasWebConfig),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get hasWindowsConfig => (super.noSuchMethod(
+        Invocation.getter(#hasWindowsConfig),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get hasMacOSConfig => (super.noSuchMethod(
+        Invocation.getter(#hasMacOSConfig),
         returnValue: false,
       ) as bool);
   @override
@@ -97,7 +111,7 @@ class MockFlutterLauncherIconsConfig extends _i1.Mock
 /// A class which mocks [WindowsConfig].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockWindowsConfig extends _i1.Mock implements _i3.WindowsConfig {
+class MockWindowsConfig extends _i1.Mock implements _i4.WindowsConfig {
   MockWindowsConfig() {
     _i1.throwOnMissingStub(this);
   }
@@ -120,7 +134,7 @@ class MockWindowsConfig extends _i1.Mock implements _i3.WindowsConfig {
 /// A class which mocks [FLILogger].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockFLILogger extends _i1.Mock implements _i4.FLILogger {
+class MockFLILogger extends _i1.Mock implements _i5.FLILogger {
   MockFLILogger() {
     _i1.throwOnMissingStub(this);
   }
